% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{data}
\name{bioclim_gfdl-esm4_historical_1995_landonly}
\alias{bioclim_gfdl-esm4_historical_1995_landonly}
\title{Global bioclimatic data for 30-yr period centered around 1995 acquired from 
historical GFDL-ESM4 model data}
\format{
\code{data.frame}
}
\description{
data.frame with global bioclimatic for 1995 according to 
GFDL-ESM4 historical model data
}
\details{
This data.frame includes the global bioclimatic data for 1995 calculated 
from 30-yr daily simulated GFDL-ESM4 data centered around 1995
}
