% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{data}
\name{bioclim_gfdl-esm4_ssp126_2080_landonly}
\alias{bioclim_gfdl-esm4_ssp126_2080_landonly}
\title{Global bioclimatic data for 30-yr period centered around 2080 
acquired from GFDL-ESM4 model data under SSP126}
\format{
\code{data.frame}
}
\description{
data.frame with global bioclimatic for 2080 according to GFDL-ESM4 model 
data under SSP126.
}
\details{
This data.frame includes the global bioclimatic data for 2080 calculated 
from 30-yr daily simulated GFDL-ESM4 data centered around 2080 based on the SSP126.
}
