% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{data}
\name{bioclim_gswp3-w5e5_obsclim_2000_landonly}
\alias{bioclim_gswp3-w5e5_obsclim_2000_landonly}
\title{Global bioclimatic data for 30-yr period centered around 2000 acquired from 
observed GSWP3-W5E5 re-analysis climate data}
\format{
\code{data.frame}
}
\description{
data.frame with global bioclimatic for 2000 according to 
observed GSWP3-W5E5 re-analysis climate data
}
\details{
This data.frame includes the global bioclimatic data for 2000 calculated 
from 30-yr daily observed GSWP3-W5E5 re-analysis climate data centered around 2000
}
