% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{data}
\name{bioclim_hadgem2-es_rcp26_2080_landonly}
\alias{bioclim_hadgem2-es_rcp26_2080_landonly}
\title{Global bioclimatic data for 30-yr period centered around 2080 acquired from HadGEM2-ES model data under RCP2.6}
\format{
\code{data.frame}
}
\description{
data.frame with global bioclimatic for 2080 according to HadGEM2-ES model data under RCP2.6
}
\details{
This data.frame includes the global bioclimatic data for 2080 calculated from 30-yr daily simulated
HadGEM2-ES data centered around 2080 based on the representative concentration pathway RCP2.6
}
