% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{data}
\name{bioclim_ipsl-cm5a-lr_rcp85_2050_landonly}
\alias{bioclim_ipsl-cm5a-lr_rcp85_2050_landonly}
\title{Global bioclimatic data for 30-yr period centered around 2050 acquired from IPSL-CM5A-LR model data under RCP8.5}
\format{
\code{data.frame}
}
\description{
data.frame with global bioclimatic for 2050 according to IPSL-CM5A-LR model data under RCP8.5
}
\details{
This data.frame includes the global bioclimatic data for 2050 calculated from 30-yr daily simulated
IPSL-CM5A-LR data centered around 2050 based on the representative concentration pathway RCP8.5
}
