% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{data}
\name{bioclim_ipsl-cm6a-lr_historical_1995_landonly}
\alias{bioclim_ipsl-cm6a-lr_historical_1995_landonly}
\title{Global bioclimatic data for 30-yr period centered around 1995 acquired from 
historical IPSL-CM6A-LR model data}
\format{
\code{data.frame}
}
\description{
data.frame with global bioclimatic for 1995 according to 
IPSL-CM6A-LR historical model data
}
\details{
This data.frame includes the global bioclimatic data for 1995 calculated 
from 30-yr daily simulated IPSL-CM6A-LR data centered around 1995
}
