% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{data}
\name{bioclim_ipsl-cm6a-lr_historical_2000_landonly}
\alias{bioclim_ipsl-cm6a-lr_historical_2000_landonly}
\title{Global bioclimatic data for 30-yr period centered around 2000 acquired from 
historical IPSL-CM6A-LR model data}
\format{
\code{data.frame}
}
\description{
data.frame with global bioclimatic for 2000 according to 
IPSL-CM6A-LR historical model data
}
\details{
This data.frame includes the global bioclimatic data for 2000 calculated 
from 30-yr daily simulated IPSL-CM6A-LR data centered around 2000
}
