% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{data}
\name{bioclim_ipsl-cm6a-lr_ssp370_2080_landonly}
\alias{bioclim_ipsl-cm6a-lr_ssp370_2080_landonly}
\title{Global bioclimatic data for 30-yr period centered around 2080 
acquired from IPSL-CM6A-LR model data under SSP370}
\format{
\code{data.frame}
}
\description{
data.frame with global bioclimatic for 2080 according to IPSL-CM6A-LR model 
data under SSP370.
}
\details{
This data.frame includes the global bioclimatic data for 2080 calculated 
from 30-yr daily simulated IPSL-CM6A-LR data centered around 2080 based on the SSP370.
}
