% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{data}
\name{bioclim_ipsl-cm6a-lr_ssp585_2050_landonly}
\alias{bioclim_ipsl-cm6a-lr_ssp585_2050_landonly}
\title{Global bioclimatic data for 30-yr period centered around 2050 
acquired from IPSL-CM6A-LR model data under SSP585}
\format{
\code{data.frame}
}
\description{
data.frame with global bioclimatic for 2050 according to IPSL-CM6A-LR model 
data under SSP585.
}
\details{
This data.frame includes the global bioclimatic data for 2050 calculated 
from 30-yr daily simulated IPSL-CM6A-LR data centered around 2050 based on the SSP585.
}
