% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{data}
\name{bioclim_miroc5_rcp60_2080_landonly}
\alias{bioclim_miroc5_rcp60_2080_landonly}
\title{Global bioclimatic data for 30-yr period centered around 2080 acquired from MIROC5 model data under RCP6.0}
\format{
\code{data.frame}
}
\description{
data.frame with global bioclimatic for 2080 according to MIROC5 model data under RCP6.0
}
\details{
This data.frame includes the global bioclimatic data for 2080 calculated from 30-yr daily simulated
MIROC5 data centered around 2080 based on the representative concentration pathway RCP6.0
}
