% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{data}
\name{bioclim_mpi-esm1-2-hr_historical_2000_landonly}
\alias{bioclim_mpi-esm1-2-hr_historical_2000_landonly}
\title{Global bioclimatic data for 30-yr period centered around 2000 acquired from 
historical MPI-ESM1-2-HR model data}
\format{
\code{data.frame}
}
\description{
data.frame with global bioclimatic for 2000 according to 
MPI-ESM1-2-HR historical model data
}
\details{
This data.frame includes the global bioclimatic data for 2000 calculated 
from 30-yr daily simulated MPI-ESM1-2-HR data centered around 2000
}
