% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{data}
\name{bioclim_mri-esm2-0_historical_1995_landonly}
\alias{bioclim_mri-esm2-0_historical_1995_landonly}
\title{Global bioclimatic data for 30-yr period centered around 1995 acquired from 
historical MRI-ESM2-0 model data}
\format{
\code{data.frame}
}
\description{
data.frame with global bioclimatic for 1995 according to 
MRI-ESM2-0 historical model data
}
\details{
This data.frame includes the global bioclimatic data for 1995 calculated 
from 30-yr daily simulated MRI-ESM2-0 data centered around 1995
}
