% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{data}
\name{bioclim_mri-esm2-0_ssp126_2050_landonly}
\alias{bioclim_mri-esm2-0_ssp126_2050_landonly}
\title{Global bioclimatic data for 30-yr period centered around 2050 
acquired from MRI-ESM2-0 model data under SSP126}
\format{
\code{data.frame}
}
\description{
data.frame with global bioclimatic for 2050 according to MRI-ESM2-0 model 
data under SSP126
}
\details{
This data.frame includes the global bioclimatic data for 2050 calculated 
from 30-yr daily simulated MRI-ESM2-0 data centered around 2050 based on the SSP126.
}
