% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{data}
\name{bioclim_ukesm1-0-ll_historical_2000_landonly}
\alias{bioclim_ukesm1-0-ll_historical_2000_landonly}
\title{Global bioclimatic data for 30-yr period centered around 2000 acquired from 
historical UKESM1-0-LL model data}
\format{
\code{data.frame}
}
\description{
data.frame with global bioclimatic for 2000 according to 
UKESM1-0-LL historical model data
}
\details{
This data.frame includes the global bioclimatic data for 2000 calculated 
from 30-yr daily simulated UKESM1-0-LL data centered around 2000
}
