% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{data}
\name{bioclim_ukesm1-0-ll_ssp126_2050_landonly}
\alias{bioclim_ukesm1-0-ll_ssp126_2050_landonly}
\title{Global bioclimatic data for 30-yr period centered around 2050 
acquired from UKESM1-0-LL model data under SSP126}
\format{
\code{data.frame}
}
\description{
data.frame with global bioclimatic for 2050 according to UKESM1-0-LL model 
data under SSP126
}
\details{
This data.frame includes the global bioclimatic data for 2050 calculated 
from 30-yr daily simulated UKESM1-0-LL data centered around 2050 based on the SSP126.
}
