% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{data}
\name{delta_runmean31_tas}
\alias{delta_runmean31_tas}
\title{Years of global mean temperature thresholds}
\format{
\code{data.frame}
}
\description{
data.frame with years of global mean temperature thresholds according to ISIMIP2b climate models
}
\details{
This data.frame depicts the expert year in which the 31-year running mean of 
global mean temperature crosses the given thresholds for different models and rcps.
}
