% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{data}
\name{landseamask_generic}
\alias{landseamask_generic}
\title{Land sea mask of ISIMIP data}
\format{
\code{RasterLayer}
}
\description{
RasterLayer with land sea mask of ISIMIP data
}
\details{
This RasterLayer depicts the global land sea mask used for gridded ISIMIP data at a resolution of 0.5 degree.
}
