% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{data}
\name{landuse-totals_2005soc}
\alias{landuse-totals_2005soc}
\title{Total landuse of 2005}
\format{
\code{data.frame}
}
\description{
data.frame with percentage cover of each totals landuse class for 2005
}
\details{
This data.frame depicts the total landuse coverage for 2005 under the 2005soc scenario.
}
