% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{data}
\name{landuse-totals_rcp26_hadgem2-es_1995}
\alias{landuse-totals_rcp26_hadgem2-es_1995}
\title{Global total landuse for 30-yr period centered around 1995 derived from HadGEM2-ES model data under RCP2.6}
\format{
\code{data.frame}
}
\description{
data.framewith percentage cover of each totals landuse class for 1995 
according to HadGEM2-ES model data under RCP2.6
}
\details{
This data.frame includes total landuse coverage for 1995 calculated from 30-yr yearly simulated
HadGEM2-ES data centered around 1995 based on the representative concentration pathway RCP2.6
}
