% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{data}
\name{landuse-totals_rcp26_ipsl-cm5a-lr_2080}
\alias{landuse-totals_rcp26_ipsl-cm5a-lr_2080}
\title{Global total landuse for 30-yr period centered around 2080 derived from IPSL-CM5A-LR model data under RCP2.6}
\format{
\code{data.frame}
}
\description{
data.framewith percentage cover of each totals landuse class for 2080 
according to IPSL-CM5A-LR model data under RCP2.6
}
\details{
This data.frame includes total landuse coverage for 2080 calculated from 30-yr yearly simulated
IPSL-CM5A-LR data centered around 2080 based on the representative concentration pathway RCP2.6
}
