% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{data}
\name{landuse-totals_rcp60_hadgem2-es_2080}
\alias{landuse-totals_rcp60_hadgem2-es_2080}
\title{Global total landuse for 30-yr period centered around 2080 derived from HadGEM2-ES model data under RCP6.0}
\format{
\code{data.frame}
}
\description{
data.framewith percentage cover of each totals landuse class for 2080 
according to HadGEM2-ES model data under RCP6.0
}
\details{
This data.frame includes total landuse coverage for 2080 calculated from 30-yr yearly simulated
HadGEM2-ES data centered around 2080 based on the representative concentration pathway RCP6.0
}
