% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{data}
\name{landuse-totals_rcp60_ipsl-cm5a-lr_1995}
\alias{landuse-totals_rcp60_ipsl-cm5a-lr_1995}
\title{Global total landuse for 30-yr period centered around 1995 derived from IPSL-CM5A-LR model data under RCP6.0}
\format{
\code{data.frame}
}
\description{
data.framewith percentage cover of each totals landuse class for 1995 
according to IPSL-CM5A-LR model data under RCP2.6
}
\details{
This data.frame includes total landuse coverage for 1995 calculated from 30-yr yearly simulated
IPSL-CM5A-LR data centered around 1995 based on the representative concentration pathway RCP2.6
}
