% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listISIMIP.R
\name{listISIMIP}
\alias{listISIMIP}
\title{List ISIMIP Environmental Data files}
\usage{
listISIMIP(
  path = getwd(),
  version = "ISIMIP2b",
  type = "GCM",
  extent = "landonly",
  model = "IPSL-CM5A-LR",
  scenario = "rcp26",
  startyear = 2006,
  endyear = 2010,
  var = "tas"
)
}
\arguments{
\item{path}{Path to files}

\item{version}{Version of ISIMIP data, default is ISIMIP2b, can be one of ISIMIP2b, ISIMIP3a, ISIMIP3b}

\item{type}{One of GCM or landuse to specify which type of data you want to get}

\item{extent}{Extent of data, either global or landonly}

\item{model}{A character string specifying the Global Climate Model used}

\item{scenario}{Character specifying the scenario to use, either histsoc, 2005soc, rcp26, rcp60}

\item{startyear}{Integer specifying the startyear for data retrieval}

\item{endyear}{Integer specifying the endyear for data retrieval}

\item{var}{Character specifying the required variable name, one of tas, pr, ...}
}
\value{
list of file paths with required data
}
\description{
Creates a list with required ISIMIP2b Data files
}
\examples{
\dontrun{
listISIMIP(type="landuse", scenario="2005soc", var="urbanareas")
}
}
