% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{data}
\name{protectedareas_annual_1819_2018_landonly}
\alias{protectedareas_annual_1819_2018_landonly}
\title{Annual protected areas data at 0.5 degree}
\format{
\code{data.frame}
}
\description{
data.frame with protected areas coverage per year at 0.5 degree resolution
}
\details{
This data.frame depicts the annual percentage coverage of protected areas for each grid cell 
at a spatial resolution of 0.5 degree for the years 1819 to 2018.
}
