% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{data}
\name{protectedareas_iucn_cat_2018_landonly}
\alias{protectedareas_iucn_cat_2018_landonly}
\title{Protected area coverage by IUCN category at 0.5 degree}
\format{
\code{data.frame}
}
\description{
data.frame with protected area coverage for each IUCN category in 2018 at 0.5 degree resolution
}
\details{
This data.frame depicts the percentage coverage of protected areas per IUCN category for each grid cell 
at a spatial resolution of 0.5 degree for the year 2018.
}
