% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readISIMIP.R
\name{readISIMIP}
\alias{readISIMIP}
\title{Read ISIMIP Environmental Data from NetCDF Files}
\usage{
readISIMIP(
  path = getwd(),
  version = "ISIMIP2b",
  type = "GCM",
  extent = "landonly",
  model = "IPSL-CM5A-LR",
  scenario = "rcp26",
  startyear = 2006,
  endyear = 2010,
  var = "tas"
)
}
\arguments{
\item{path}{Path to ISIMIP files}

\item{version}{Version of ISIMIP data, default is ISIMIP2b, can be one of ISIMIP2b, ISIMIP3a, ISIMIP3b}

\item{type}{One of EWEMBI, GCM, landuse or population, to specify which type of data you want to use.}

\item{extent}{Extent of data, either global or landonly}

\item{model}{A character string specifying the Global Climate Model used}

\item{scenario}{A character string specifying the scenario, either histsoc, 2005soc, rcp26, rcp60 or ssp2soc}

\item{startyear}{Startyear of timeperiod for required data}

\item{endyear}{Endyear of timeperiod for required data}

\item{var}{A character string specifying the variable to use (tas, tmin, tmax, huss, hurss, pr, ...)}
}
\value{
raster stack or list of raster stacks with all the data
}
\description{
Read and pre-process ISIMIP Data
}
\examples{
\dontrun{
readISIMIP(type="landuse", scenario="rcp26", var="urbanareas", startyear=2010, endyear=2020)
}
}
