% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{package}
\name{rISIMIP-package}
\alias{rISIMIP-package}
\alias{rISIMIPpackage}
\title{R Package for handling ISIMIP data}
\description{
R Package for handling ISIMIP data
}
\details{
Reading and processing ISIMIP NetCDF files
}
\author{
RS-eco
}
\keyword{package}
