% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{data}
\name{runmean31_tas}
\alias{runmean31_tas}
\title{31-year runing mean of global mean temperature}
\format{
\code{data.frame}
}
\description{
Dataframe with 31-year running mean past and future simulated global mean temperature
}
\details{
This data.frame depicts the 31-year running mean of global mean temperature 
for different models and rcps.
}
