% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rISIMIP-package.R
\docType{data}
\name{yearmean_tas}
\alias{yearmean_tas}
\title{Annual global mean temperature}
\format{
\code{data.frame}
}
\description{
Dataframe with past and future simulated global annual mean temperature
}
\details{
This data.frame depicts the annual global mean temperature 
for past and future time periods based on different model algorthims and rcps.
}
